+++ src/backend/joanie/core/enums.py
@@ -1,8 +1,19 @@
 """
 Core application enums declaration
 """
+from django.conf import global_settings, settings
 from django.utils.translation import gettext_lazy as _
 from django.utils.translation import pgettext_lazy
+
+# Django sets `LANGUAGES` by default with all supported languages. We can use it for example for
+# the choice of languages on the course run which should not be limited to the few languages
+# active in the app.
+# pylint: disable=no-member
+ALL_LANGUAGES = getattr(
+    settings,
+    "ALL_LANGUAGES",
+    [(language, _(name)) for language, name in global_settings.LANGUAGES],
+)
 
 PRODUCT_TYPE_CREDENTIAL = "credential"
 PRODUCT_TYPE_ENROLLMENT = "enrollment"